+++ packages/runtime/runtime-test-utils/src/mockStorage.ts
@@ -6,6 +6,7 @@
 import assert from "assert";
 import { IBlob, ITree } from "@fluidframework/protocol-definitions";
 import { IObjectStorageService } from "@fluidframework/component-runtime-definitions";
+import { listBlobsAtTreePath } from "@fluidframework/runtime-utils";
 
 /**
  * Mock implementation of IObjectStorageService based on ITree input.
@@ -46,28 +47,6 @@ export class MockStorage implements IObjectStorageService {
     }
 
     public async list(path: string): Promise<string[]> {
-        const pathParts = path.split("/").filter((v)=>v !== "");
-        let tree = this.tree;
-        while (tree?.entries !== undefined && pathParts.length > 0) {
-            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-            const part = pathParts.shift()!;
-            const index = tree.entries.findIndex((value)=>{
-                if (value.type === "Tree" && value.path === part) {
-                    return true;
-                } else {
-                    return false;
-                }
-            });
-            if (index === -1) {
-                tree = undefined;
-            } else {
-                const treeEntry = tree.entries[index];
-                tree = treeEntry.value as ITree;
-            }
-        }
-        if (tree?.entries === undefined || pathParts.length !== 0) {
-            throw new Error("path does not exist");
-        }
-        return tree.entries.filter((e)=>e.type === "Blob").map((e)=>e.path);
+        return listBlobsAtTreePath(this.tree, path);
     }
 }
