+++ packages/runtime/runtime-utils/src/index.ts
@@ -7,3 +7,4 @@ export * from "./serializer";
 export * from "./summaryTracker";
 export * from "./utils";
 export * from "./objectstoragepartition";
+export * from "./objectstorageutils";
